error handling

+++ controller/tnt-last.js
@@ -2,6 +2,7 @@ var express = require('express');
 var _ = require('lodash');
 var request = require('request');
 var cheerio = require('cheerio');
+var Regex = require("regex");
 
 var htmlAPI = 'http://www.tntexpress.com.au/InterAction/ASPs/CnmHxAS.asp?';
 var xmlAPI = 'http://sttrackandtrace.startrack.com.au/Consignment/GetConsignmentEventsByConsignmentGuid/';
@@ -20,43 +21,40 @@ module.exports.getTNT = function (con, cb) {
 			$('.f2').each(function(i, elem) {
 				conArray[i] = $(this).text()
 			})
+			var pattern = new Regex(/(No data)/i)
+
+			// console.log(pattern.test(conArray[0]))
+
+			if (!conArray || conArray[0].split(' ')[0] == 'No') {
+				cb('No Data for consignment number ' + con)
+			} else {
+
+				var consignment = _
+	                .chain(conArray)
+	                .drop()
+	                .chunk(4)
+	                .last()
+	                .value()
+ 
+	                // console.log(consignment)
+
+		                sendObj = {
+		                  conNum: '',
+		                  status: '',
+		                  date: '',
+		                  time: '',
+		                  depot: ''
+		                }
+
+	                	sendObj.conNum = conArray[0].split(' ')[2]
+	                    sendObj.status = consignment[0]
+	                    sendObj.date = consignment[1]
+	                    sendObj.time = consignment[2]
+	                    sendObj.depot = consignment[3]
 
 
-
-			var consignment = _
-                .chain(conArray)
-                .drop()
-                .chunk(4)
-                .last()
-                .value()
-
-
-                // console.log(consignment)
-
-                // for (var i = 0; i < consignment.length; i++) {
-
-	                sendObj = {
-	                  conNum: '',
-	                  status: '',
-	                  date: '',
-	                  time: '',
-	                  depot: ''
-	                }
-
-                	sendObj.conNum = conArray[0].split(' ')[2]
-                    sendObj.status = consignment[0]
-                    sendObj.date = consignment[1]
-                    sendObj.time = consignment[2]
-                    sendObj.depot = consignment[3]
-
-                    // console.log(sendObj)
-                    // dataArray.push(sendObj)
-                // }
-
-
-
-
-			cb(sendObj)
+				cb(sendObj)
+			}
 		}
 	})
 }
